+++ src/interfaces/auth-checker.ts
@@ -4,3 +4,5 @@ export type AuthChecker<ContextType = {}> = (
   actionData: ActionData<ContextType>,
   roles: string[],
 ) => boolean | Promise<boolean>;
+
+export type AuthMode = "error" | "null";